requirements.txt
ftfy==6.1.1
google-api-python-client==1.7.2
python-dotenv==1.0.0
underthesea==6.7.0
youtube-transcript-api==0.6.1